updated ES6 and index files
