+++ index.html
@@ -87,7 +87,8 @@
       </div>
     </footer>
     <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="/javascript/posts.js"></script>
-    <script src="/javascript/createPost.js"></script>
+    <script type="module" src="/javascript/api/actions/index.js"></script>
+    <script type="module" src="/javascript/api/actions/functions/createPost.js"></script>
+    <script src="/javascript/api/actions/functions/filterpost.js"></script>
   </body>
 </html>
