+++ javascript/api/actions/urls/url.js
@@ -1,0 +1,15 @@
+const baseUrl = "https://nf-api.onrender.com/api/v1/social/";
+const postUrl = baseUrl + "posts";
+const authUrl = baseUrl + "auth/";
+const loginUrl = authUrl + "login";
+const registerUrl = authUrl + "register";
+
+// let params = "limit=10&_author=true";
+// let url = `https://nf-api.onrender.com/api/v1/social/posts?${params}`;
+// const url2 = "https://nf-api.onrender.com/api/v1/social/posts?_author=true";
+// const url = "https://nf-api.onrender.com/api/v1/social/posts?limit=5000&_author=true";
+// const url2 = "https://nf-api.onrender.com/api/v1/social/posts?_author=true";
+// const url = "https://nf-api.onrender.com/api/v1/social/auth/login";
+// const url = "https://nf-api.onrender.com/api/v1/social/auth/register";
+
+export { baseUrl, postUrl, authUrl, loginUrl, registerUrl };
