+++ pages/login.html
@@ -61,7 +61,7 @@
           </div>
           <button type="submit" class="btn btn-primary mb-3">Submit</button>
         </form>
-        <a href="/register.html">Dont have an Account? Click here to make one</a>
+        <a href="/pages/register.html">Dont have an Account? Click here to make one</a>
       </div>
     </div>
 
@@ -73,6 +73,6 @@
       </div>
     </footer>
     <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="/javascript/login.js"></script>
+    <script type="module" src="/javascript/api/auth/login.js"></script>
   </body>
 </html>
