+++ pages/profile.html
@@ -117,6 +117,6 @@
       </div>
     </footer>
     <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="/javascript/profile.js"></script>
+    <script type="module" src="/javascript/api/actions/profile.js"></script>
   </body>
 </html>
