+++ pages/register.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Gettr | Register</title>
-    <link rel="stylesheet" href="css/main.css" />
+    <link rel="stylesheet" href="../css/main.css" />
   </head>
   <body class="">
     <!-- Navigation -->
@@ -65,7 +65,7 @@
           </div>
           <button type="submit" class="btn btn-primary mb-3" id="register-button">Register</button>
         </form>
-        <a href="/login.html">Have an Account? Click here to Login</a>
+        <a href="/pages/login.html">Have an Account? Click here to Login</a>
       </div>
     </div>
 
@@ -77,6 +77,6 @@
       </div>
     </footer>
     <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="/javascript/register.js"></script>
+    <script type="module" src="/javascript/api/auth/register.js"></script>
   </body>
 </html>
