+++ pages/specific.html
@@ -72,6 +72,6 @@
       </div>
     </footer>
     <script src="node_modules/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="/javascript/specific.js"></script>
+    <script type="module" src="/javascript//api/actions/specific.js"></script>
   </body>
 </html>
